moving around tuning code
