+++ src/torchpme/__init__.py
@@ -1,7 +1,7 @@
 import contextlib
 
 from .calculators.base import Calculator
-from .calculators.ewald import EwaldCalculator, tune_ewald
+from .calculators.ewald import EwaldCalculator
 from .calculators.pme import PMECalculator
 from .lib.potentials import CoulombPotential, InversePowerLawPotential
 
@@ -14,6 +14,5 @@ __all__ = [
     "PMECalculator",
     "CoulombPotential",
     "InversePowerLawPotential",
-    "tune_ewald",
 ]
 __version__ = "0.0.0-dev"
