+++ tests/calculators/test_base.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from torchpme.calculators import Calculator, estimate_smearing, get_cscl_data
+from torchpme.calculators import Calculator
 from torchpme.lib.potentials import CoulombPotential
 
 # Define some example parameters
@@ -20,14 +20,6 @@ class CalculatorTest(Calculator):
         super().__init__(
             potential=CoulombPotential(smearing=None, exclusion_radius=None)
         )
-
-
-def test_cscl_data():
-    data = get_cscl_data()
-
-    assert len(data) == 7
-    for el in data:
-        assert type(el) is torch.Tensor
 
 
 def test_compute_output_shapes():
@@ -249,12 +241,3 @@ def test_invalid_device_neighbor_distances():
             neighbor_indices=torch.ones((10, 2), dtype=DTYPE, device=DEVICE),
             neighbor_distances=torch.ones((10), dtype=DTYPE, device="meta"),
         )
-
-
-def test_no_cell():
-    match = (
-        "provided `cell` has a determinant of 0 and therefore is not valid for "
-        "periodic calculation"
-    )
-    with pytest.raises(ValueError, match=match):
-        estimate_smearing(cell=torch.zeros(3, 3))
